requirements.txt
numpy~=1.18.5
PyYAML~=5.3.1
scikit-image~=0.17.2
tensorflow>=2.3.0
pycocotools~=2.0.2
scipy~=1.4.1